Add school server helpers and UI features
- Introduce apps/www/lib/school/server.ts with cached helpers:
  getSchoolInfoBySlug, getSchoolRouteSnapshot, getClassRouteSnapshot
- Use next/navigation notFound and move route 404 into a shared
  not-found.tsx for the school route family
- Add class creation form schema and defaults (header-add-schema)
- Add MaterialGroupEditorDialog and refactor materials new/item to use
  it
- Add invite-data helper for class invite role mapping
- Remove direct fetchQuery/import duplication and centralize server
  error capture where appropriate
