+++ apps/www/app/[locale]/(app)/school/(authenticated)/[slug]/(main)/classes/[id]/layout.tsx
@@ -1,17 +1,13 @@
-import { captureServerException } from "@repo/analytics/posthog/server";
-import { api } from "@repo/backend/convex/_generated/api";
 import type { Id } from "@repo/backend/convex/_generated/dataModel";
-import { ErrorBoundary } from "@repo/design-system/components/ui/error-boundary";
-import { fetchQuery } from "convex/nextjs";
+import { notFound } from "next/navigation";
 import { use } from "react";
 import { SchoolClassesForumPostSheet } from "@/components/school/classes/forum/post-sheet";
 import { SchoolClassesHeaderInfo } from "@/components/school/classes/info";
 import { SchoolClassesJoinForm } from "@/components/school/classes/join-form";
 import { SchoolClassesTabs } from "@/components/school/classes/tabs";
-import { SchoolNotFound } from "@/components/school/not-found";
-import { getToken } from "@/lib/auth/server";
 import { ClassContextProvider } from "@/lib/context/use-class";
 import { ForumContextProvider } from "@/lib/context/use-forum";
+import { getClassRouteSnapshot } from "@/lib/school/server";
 
 /** Bind the resolved class route snapshot to the class subtree. */
 export default function Layout(
@@ -36,47 +32,30 @@ async function ClassRouteBoundary({
   children: React.ReactNode;
   classId: Id<"schoolClasses">;
 }) {
-  const token = await getToken();
+  const value = await getClassRouteSnapshot({ classId });
 
-  if (!token) {
-    return <SchoolNotFound />;
+  if (!value) {
+    notFound();
   }
 
-  try {
-    const value = await fetchQuery(
-      api.classes.queries.getClassRoute,
-      { classId },
-      { token }
-    );
-
-    if (value.kind === "joinRequired") {
-      return (
-        <SchoolClassesJoinForm
-          classId={value.class._id}
-          visibility={value.class.visibility}
-        />
-      );
-    }
-
+  if (value.kind === "joinRequired") {
     return (
-      <ErrorBoundary fallback={<SchoolNotFound />}>
-        <ClassContextProvider value={value}>
-          <SchoolClassesHeaderInfo />
-          <SchoolClassesTabs />
-
-          <ForumContextProvider key={classId}>
-            <SchoolClassesForumPostSheet />
-            {children}
-          </ForumContextProvider>
-        </ClassContextProvider>
-      </ErrorBoundary>
+      <SchoolClassesJoinForm
+        classId={value.class._id}
+        visibility={value.class.visibility}
+      />
     );
-  } catch (error) {
-    await captureServerException(error, undefined, {
-      classId,
-      source: "school-class-route-boundary",
-    });
-
-    return <SchoolNotFound />;
   }
+
+  return (
+    <ClassContextProvider value={value}>
+      <SchoolClassesHeaderInfo />
+      <SchoolClassesTabs />
+
+      <ForumContextProvider key={classId}>
+        <SchoolClassesForumPostSheet />
+        {children}
+      </ForumContextProvider>
+    </ClassContextProvider>
+  );
 }
