+++ apps/www/app/[locale]/(app)/school/not-found.tsx
@@ -1,0 +1,6 @@
+import { SchoolNotFound } from "@/components/school/not-found";
+
+/** Render the shared Nakafa School 404 screen for the school route family. */
+export default function NotFound() {
+  return <SchoolNotFound />;
+}
